Rename ncm component for consistency.
