+++ BUILD.bazel
@@ -118,8 +118,8 @@ exports_files(glob(
 # gazelle:exclude comp/netflow/server
 # gazelle:exclude comp/netflow/testutil
 # gazelle:exclude comp/netflow/topn
-# gazelle:exclude comp/networkdeviceconfig/fx
-# gazelle:exclude comp/networkdeviceconfig/impl
+# gazelle:exclude comp/networkconfigmanagement
+# gazelle:exclude comp/networkpath
 # gazelle:exclude comp/notableevents/fx
 # gazelle:exclude comp/notableevents/impl
 # gazelle:exclude comp/offlinereporter
