+++ comp/networkconfigmanagement/impl/config.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025-present Datadog, Inc.
 
-package networkdeviceconfigimpl
+package networkconfigmanagementimpl
 
 import (
 	"github.com/DataDog/datadog-agent/comp/core/config"
