+++ comp/networkconfigmanagement/impl/remote.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025-present Datadog, Inc.
 
-package networkdeviceconfigimpl
+package networkconfigmanagementimpl
 
 import "golang.org/x/crypto/ssh"
