--- comp/networkdeviceconfig/def/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "def",
-    srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/networkdeviceconfig/def",
-    visibility = ["//visibility:public"],
-)
