--- comp/networkdeviceconfig/mock/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "mock",
-    srcs = ["mock.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/networkdeviceconfig/mock",
-    visibility = ["//visibility:public"],
-)
